sync

+++ app/code/Magento/Paypal/Model/Cart.php
@@ -18,8 +18,6 @@
  * versions in the future. If you wish to customize Magento for your
  * needs please refer to http://www.magentocommerce.com for more information.
  *
- * @category    Magento
- * @package     Magento_Paypal
  * @copyright   Copyright (c) 2014 X.commerce, Inc. (http://www.magentocommerce.com)
  * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
@@ -56,9 +54,7 @@ class Cart extends \Magento\Payment\Model\Cart
                 $subtotal -= $this->getDiscount();
             }
 
-            return array(
-                self::AMOUNT_SUBTOTAL => $subtotal
-            );
+            return array(self::AMOUNT_SUBTOTAL => $subtotal);
         }
 
         return $this->_amounts;
@@ -153,14 +149,17 @@ class Cart extends \Magento\Payment\Model\Cart
 
             // aggregate item price if item qty * price does not match row total
             $itemBaseRowTotal = $item->getOriginalItem()->getBaseRowTotal();
-            if (($amount * $qty) != $itemBaseRowTotal) {
-                $amount = (float)$itemBaseRowTotal;
+            if ($amount * $qty != $itemBaseRowTotal) {
+                $amount = (double)$itemBaseRowTotal;
                 $subAggregatedLabel = ' x' . $qty;
                 $qty = 1;
             }
 
-            $this->_salesModelItems[] = $this->_createItemFromData($item->getName() . $subAggregatedLabel, $qty,
-                $amount);
+            $this->_salesModelItems[] = $this->_createItemFromData(
+                $item->getName() . $subAggregatedLabel,
+                $qty,
+                $amount
+            );
         }
 
         $this->addSubtotal($this->_salesModel->getBaseSubtotal());
@@ -191,10 +190,11 @@ class Cart extends \Magento\Payment\Model\Cart
      * @param \Magento\Payment\Model\Cart\SalesModel\SalesModelInterface $salesEntity
      * @return void
      */
-    protected function _applyHiddenTaxWorkaround(\Magento\Payment\Model\Cart\SalesModel\SalesModelInterface $salesEntity)
-    {
+    protected function _applyHiddenTaxWorkaround(
+        \Magento\Payment\Model\Cart\SalesModel\SalesModelInterface $salesEntity
+    ) {
         $dataContainer = $salesEntity->getTaxContainer();
-        $this->addTax((float)$dataContainer->getBaseHiddenTaxAmount());
-        $this->addTax((float)$dataContainer->getBaseShippingHiddenTaxAmnt());
+        $this->addTax((double)$dataContainer->getBaseHiddenTaxAmount());
+        $this->addTax((double)$dataContainer->getBaseShippingHiddenTaxAmnt());
     }
 }
